Add intermediary ERB file to generate active sidebar classes
This commit improves page generation performance by generating the
sidebar once, including it in each page and handling active classes
based on the current page; as opposed to the sidebar being generated
with active classes by looping through the list of documents for each
page.

When the Gluegun commands are run, the _sidebar-template.erb file loops
through the list of documents once and generates the _sidebar.erb file
which contains the hardcoded sidebar with the conditionals that check
if the given sidebar item/category should have active classes. The
_sidebar.erb file is included on the beginning of the _header.erb file.