Add pattern filter to yml and check for valid links w/o opening files
- This commit adds a Rules field to the .yml file, and a Filter field
under it to exclude unwanted patterns from the markdown content
fetched from Github.
- This commit also checks if the links are valid or not without opening
the file. This change was done to improve performance by preventing the
'reveal' method from running for each document while the sidebar is
being generated.
- When the first link in the list of Documents is broken, to avoid the
wrong link to be highlighted on index.html, Gluegun now returns an error
while generating the index page and asks the user to correct it.